Main.tsxからバックグランド位置情報関係の処理をカスタムフックに分離
